day6/requirements.txt
openai>=1.0.0
termcolor
python-dotenv
